Fixed another misuse of long where int32 should have been used instead. This causes an endless silly transmit loop on 64-bit machines because the local host miscalculates the ACK value to send.

+++ tcpin.c
@@ -884,7 +884,7 @@ struct tcp *seg,
 struct mbuf **bpp,
 uint *length
 ){
-	long dupcnt,excess;
+	int32 dupcnt,excess;
 	uint len;		/* Segment length including flags */
 	char kaccept = 0;
